modularize network code
