+++ src/iad/iad.c
@@ -12,6 +12,10 @@
 #include "audio/input.h"          // for initialize_audio_input_device, DEFAULT_AI...
 #include "audio/audio_common.h"   // for PlayInputAttributes, disable_audio_input
 #include "network/network.h"        // for audio_control_server_thread, audio_input_...
+#include "network/input_server.h"        // for audio_control_server_thread, audio_input_...
+#include "network/output_server.h"        // for audio_control_server_thread, audio_input_...
+#include "network/control_server.h"        // for audio_control_server_thread, audio_input_...
+#include "network/network.h"        // for audio_control_server_thread, audio_input_...
 #include "audio/output.h"         // for disable_audio_output, ao_test_play_thread
 #include "utils/cmdline.h"        // for CmdOptions, parse_cmdline
 #include "utils/config.h"         // for config_cleanup, config_get_ai_enabled
