+++ src/iad/network/control_server.h
@@ -1,5 +1,7 @@
-#ifndef NETWORK_H
-#define NETWORK_H
+#ifndef CONTROL_SERVER_H
+#define CONTROL_SERVER_H
+
+extern char AUDIO_CONTROL_SOCKET_PATH[];
 
 #define CLIENT_QUEUED 1
 #define CLIENT_NOT_QUEUED 0
@@ -17,8 +19,7 @@
 #define RESPONSE_UNKNOWN_VARIABLE 404
 
 // Functions
-void update_socket_paths_from_config();
-char* get_variable_value(const char* variable_name);
-int set_variable_value(const char* variable_name, const char* value);
+void *audio_control_server_thread(void *arg);
+void handle_control_client(int client_sock);
 
-#endif // NETWORK_H
+#endif // CONTROL_SERVER_H
